requirement.txt
Flask==2.1.2
google-cloud-vision==3.0.2
google-cloud-speech==2.14.1
google-cloud-storage==2.9.0

